requirements_test.txt
pytest>=8.0
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
requests-mock>=1.11
